requirements-common.txt
# 공통 Python 의존성
# 데이터베이스
psycopg2-binary==2.9.9

# 검색 엔진
opensearch-py==2.4.0

# 캐시
redis==5.0.1

# 환경변수
python-dotenv==1.0.0

# 머신러닝 (공통)
numpy==1.24.4
pandas==2.1.4
scikit-learn==1.3.2
sentence-transformers==2.2.2

# HTTP 클라이언트
aiohttp==3.9.1
requests==2.31.0

# 웹 스크래핑
beautifulsoup4==4.12.2
lxml==4.9.3

# Google API
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
